Removed the dashboard service and updated the bill daos and services to load the number of bills and deleted the dashboard service
